ChatAPICall

Documentation for ChatAPICall.

ChatAPICall.Chat

ChatAPICall.Chat

ChatAPICall.ErrResp

ChatAPICall.Resp

Base.copy

Base.length

Base.pop!

Base.print

Base.print

Base.show

ChatAPICall.add!

ChatAPICall.add!

ChatAPICall.addassistant!

ChatAPICall.addsystem!

ChatAPICall.adduser!

ChatAPICall.defaultprompt

ChatAPICall.getresponse

ChatAPICall.getresponse!

ChatAPICall.proxy_off

ChatAPICall.proxy_on

ChatAPICall.proxy_status

ChatAPICall.setapikey

ChatAPICall.setbaseurl

ChatAPICall.showapikey

ChatAPICall.Chat — Type

Chat

Chat with OpenAI API.

Fields

chatlog::Vector{Dict}: The chat log, default to be Dict[].

apikey::String: The API key, default to be ChatAPICall.apikey.

ChatAPICall.Chat — Method

Chat(msg)

Initialize Chat object by a message.

By default, the apikey is set to be ChatAPICall.apikey.

ChatAPICall.ErrResp — Type

ErrResp

Error response from OpenAI API.

Fields

message::String: The error message.

type::String: The error type.

param::Union{Nothing, String}: The parameter that caused the error.

code::String: The error code.

ChatAPICall.Resp — Type

Resp

Response from OpenAI API.

Fields

total_tokens::Int: Total number of tokens in the response.

prompt_tokens::Int: Number of tokens in the prompt.

completion_tokens::Int: Number of tokens in the completion.

content::String: The completion.

id::String: The ID of the completion.

model::String: The model used to generate the completion.

Base.copy — Method

copy(chat::Chat)

Copy a chat log.

Base.length — Method

length(chat::Chat)

Get the number of messages in the chat log.

Base.pop! — Method

pop!(chat::Chat)

Pop the last message from the chat log.

Base.print — Method

print(io::IO, chat::Chat)

Print the chat log.

Base.print — Method

print(io::IO, chat::Chat)

Print the chat log.

Base.show — Method

show(io::IO, chat::Chat)

Show information of the chat log.

ChatAPICall.add! — Method

add!(chat::Chat, role::AbstractString, content::AbstractString)

Add a message to the chat log.

ChatAPICall.add! — Method

add!(chat::Chat, resp::Resp)

Add a response from OpenAI API to the chat log.

Arguments

chat::Chat: The chat log.

resp::Resp: The response from OpenAI API.

ChatAPICall.addassistant! — Method

addassistant!(chat::Chat, content::AbstractString)

Add an assistant message to the chat log.

ChatAPICall.addsystem! — Method

addsystem!(chat::Chat, content::AbstractString)

Add a system message to the chat log.

ChatAPICall.adduser! — Method

adduser!(chat::Chat, content::AbstractString)

Add a user message to the chat log.

ChatAPICall.defaultprompt — Method

defaultprompt(msg)

Default prompt for Chat(msg:AbstractString).

ChatAPICall.getresponse! — Method

getresponse!( chat::Chat
            , maxrequests::Int=1
            , compact::Bool=true
            , model::AbstractString="gpt-3.5-turbo"
            , options...)::Resp

Get a response from OpenAI API and add it to the chat log.

ChatAPICall.getresponse — Method

getresponse( chat::Chat
           , maxrequests::Int=1
           , compact::Bool=true
           , model::AbstractString="gpt-3.5-turbo"
           , options...)::Resp

Get a response from OpenAI API.

Arguments

chat::Chat: The chat log.

maxrequests::Int: The maximum number of requests to OpenAI API.

compact::Bool: Whether to compact the response.

model::AbstractString: The model to use.

**options: Other options to pass to ChatAPICall.request.

ChatAPICall.proxy_off — Method

proxy_off()

Turn off proxy.

ChatAPICall.proxy_on — Method

proxy_on(http::Union{Nothing, AbstractString}, https::Union{Nothing, AbstractString})

Turn on proxy.

ChatAPICall.proxy_status — Method

proxy_status()

Get proxy status.

ChatAPICall.setapikey — Method

setapikey(key::AbstractString)

Set the OpenAI API key.

ChatAPICall.setbaseurl — Method

setbaseurl(url::AbstractString)

Set the proxy URL of the OpenAI API.

ChatAPICall.showapikey — Method

showapikey()

Show the OpenAI API key.